# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TontineChain API",
        "description": "Documentation de l'API TontineChain pour le Hackathon MIABE 2026.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur Local"
        },
        {
            "url": "https://tonnine-benin-backend.onrender.com",
            "description": "Serveur de Production"
        }
    ],
    "paths": {
        "/api/v1/ai/chat": {
            "post": {
                "tags": [
                    "IA"
                ],
                "summary": "Converser avec YAO (Assistant Intelligent)",
                "operationId": "0db37d50f93bd2808dd6e4c4dbbbfa04",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Quel est mon solde ?"
                                    },
                                    "locale": {
                                        "type": "string",
                                        "example": "fr"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse de l'assistant AI",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "assistant": {
                                            "type": "string",
                                            "example": "YAO"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ton solde est de 50 000 FCFA"
                                        },
                                        "audio_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "demo_notice": {
                                            "properties": {
                                                "is_simulation": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "MULTILINGUAL AI message"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}/bid": {
            "post": {
                "tags": [
                    "Enchères"
                ],
                "summary": "Soumettre une offre d'enchère pour le tour actuel",
                "operationId": "da93214fe7b1e371b03ccb3f2acd89f4",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "discount_amount": {
                                        "description": "Montant que vous êtes prêt à laisser au groupe",
                                        "type": "number",
                                        "example": 5000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Offre enregistrée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}/bids": {
            "get": {
                "tags": [
                    "Enchères"
                ],
                "summary": "Voir les offres du cycle actuel",
                "operationId": "33c1f7164d2d0713963ede4e33939558",
                "responses": {
                    "200": {
                        "description": "Liste des offres"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/contributions/pending": {
            "get": {
                "tags": [
                    "Cotisations"
                ],
                "summary": "Mes cotisations en attente",
                "operationId": "5ea4484def1de08b71c8b81c8b7c4692",
                "responses": {
                    "200": {
                        "description": "Liste des cotisations"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/contributions/{contribution}/pay": {
            "post": {
                "tags": [
                    "Cotisations"
                ],
                "summary": "Initier le paiement d'une cotisation",
                "operationId": "962acccd533922f22bcbc22dc864063e",
                "parameters": [
                    {
                        "name": "contribution",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lien de paiement généré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "reference": {
                                            "type": "string"
                                        },
                                        "url": {
                                            "type": "string"
                                        },
                                        "demo_notice": {
                                            "properties": {
                                                "is_simulation": {
                                                    "type": "boolean"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/groups": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Lister mes groupes",
                "operationId": "8731199f9d1e321702524c8df8c5834c",
                "responses": {
                    "200": {
                        "description": "Liste des groupes"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Créer une nouvelle tontine",
                "operationId": "0dbd5970197eb93f2a1ed934c341aeab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tontine Famille"
                                    },
                                    "contribution_amount": {
                                        "type": "number",
                                        "example": 10000
                                    },
                                    "max_members": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "enum": [
                                            "weekly",
                                            "biweekly",
                                            "monthly"
                                        ]
                                    },
                                    "payout_method": {
                                        "type": "string",
                                        "enum": [
                                            "sequential",
                                            "random",
                                            "bidding"
                                        ]
                                    },
                                    "insurance_opt_in": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Groupe créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Détails d'un groupe",
                "operationId": "cc64455d914e608b0b7f53d3b73b59bb",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du groupe"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}/start": {
            "post": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Démarrer la tontine",
                "operationId": "4716379a3472bb4db88043afb5f609dd",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tontine démarrée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "contract_address": {
                                            "type": "string"
                                        },
                                        "demo_notice": {
                                            "properties": {
                                                "is_simulation": {
                                                    "type": "boolean"
                                                },
                                                "blockchain_tx": {
                                                    "type": "string"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}/invite": {
            "post": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Inviter un membre",
                "operationId": "5a69ca01630f7b36a7e1f59f01f9694c",
                "responses": {
                    "200": {
                        "description": "Invitation envoyée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "member_analysis": {
                                            "type": "object"
                                        },
                                        "demo_notice": {
                                            "properties": {
                                                "is_simulation": {
                                                    "type": "boolean"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/groups/{group}/stats": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Statistiques & Analyse de Risque IA",
                "operationId": "3e7d5fc393168a14146671e8fc31e1ad",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stats et prédiction de risque"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}/contract": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Télécharger le contrat PDF de la tontine",
                "operationId": "903230a31f2478c27c456d8db122ff16",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier PDF du contrat"
                    },
                    "404": {
                        "description": "Contrat non disponible (tontine non démarrée)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/incidents": {
            "get": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Lister les incidents (retards, amendes)",
                "operationId": "8efda3ca532d5c4cd81590f2fb61ef24",
                "responses": {
                    "200": {
                        "description": "Liste des incidents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "incidents": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "demo_notice": {
                                            "properties": {
                                                "is_simulation": {
                                                    "type": "boolean"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/incidents/{incident}": {
            "get": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Détails d'un incident spécifique",
                "operationId": "67539457efa048e7038808859f499d6a",
                "parameters": [
                    {
                        "name": "incident",
                        "in": "path",
                        "description": "ID de l'incident",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'incident"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}/messages": {
            "get": {
                "tags": [
                    "Messagerie"
                ],
                "summary": "Liste des messages du groupe",
                "operationId": "0422780e6232111ee8898a1fe6fd5a33",
                "responses": {
                    "200": {
                        "description": "Liste des messages du groupe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "Bonjour à tous"
                                            },
                                            "user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "full_name": {
                                                        "type": "string",
                                                        "example": "Jean Dupont"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messagerie"
                ],
                "summary": "Envoyer un message dans le groupe",
                "operationId": "901d5159fe051ddda2a93ff17bc86a27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Bonjour à tous"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Bonjour à tous"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "full_name": {
                                                    "type": "string",
                                                    "example": "Jean Dupont"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/payouts": {
            "get": {
                "tags": [
                    "Payouts"
                ],
                "summary": "Lister les paiements reçus (ramassages)",
                "operationId": "03d2e7a08e16ae8df39967151be97615",
                "responses": {
                    "200": {
                        "description": "Liste des paiements du membre"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/payouts/{payout}": {
            "get": {
                "tags": [
                    "Payouts"
                ],
                "summary": "Détails d'un paiement spécifique",
                "operationId": "0b53fa1a5ff258ead22b70bafd7ff811",
                "parameters": [
                    {
                        "name": "payout",
                        "in": "path",
                        "description": "ID du paiement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du paiement"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Utiliser le token Bearer retourné par /api/v1/auth/verify-otp",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}